Move CrossPlatformStLink and make available as export
